XL/requirements.txt
torch
torchvision
numpy==1.23.3
#osgeo
pandas==1.5.0
Pillow==9.3.0
scikit_learn==1.1.3
tensorboardX==2
tqdm==4.64.1
ipykernel
